Moving encryption library to common package
